python/requirements.txt
pandas
sqlalchemy
psycopg2-binary
panel
python-dotenv
nbconvert
matplotlib
